+++ pathways/models.py
@@ -2,6 +2,18 @@ from django.db import models
 from postgres_copy import CopyManager
 
 
+class KeggInfo(models.Model):
+    """Kegg information."""
+
+    #: Kegg ID
+    kegg_id = models.CharField(max_length=16)
+    #: Kegg name
+    name = models.CharField(max_length=512)
+
+    #: Allow bulk import
+    objects = CopyManager()
+
+
 class EnsemblToKegg(models.Model):
     """Linking ensembl gene id and kegg id."""
 
@@ -12,6 +24,10 @@ class EnsemblToKegg(models.Model):
 
     #: Allow bulk import
     objects = CopyManager()
+
+    @property
+    def kegginfo(self):
+        return KeggInfo.objects.get(id=self.kegginfo_id)
 
     class Meta:
         unique_together = ["gene_id", "kegginfo_id"]
@@ -29,18 +45,10 @@ class RefseqToKegg(models.Model):
     #: Allow bulk import
     objects = CopyManager()
 
+    @property
+    def kegginfo(self):
+        return KeggInfo.objects.get(id=self.kegginfo_id)
+
     class Meta:
         unique_together = ["gene_id", "kegginfo_id"]
         indexes = [models.Index(fields=["gene_id"])]
-
-
-class KeggInfo(models.Model):
-    """Kegg information."""
-
-    #: Kegg ID
-    kegg_id = models.CharField(max_length=16)
-    #: Kegg name
-    name = models.CharField(max_length=512)
-
-    #: Allow bulk import
-    objects = CopyManager()